refactor(via): intuitive public module names
