+++ examples/file-server/src/main.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 use std::process::ExitCode;
-use via::middleware::error_boundary;
+use via::builtin::rescue;
 use via::response::File;
 use via::{BoxError, Next, Request};
 
@@ -59,11 +59,9 @@ async fn main() -> Result<ExitCode, BoxError> {
     // Create a new application.
     let mut app = via::app(());
 
-    // Include an error boundary to catch any errors that occur downstream.
-    app.include(error_boundary::map(|error| {
-        eprintln!("error: {}", error);
-        error.use_canonical_reason()
-    }));
+    // Capture errors from downstream, log them, and map them into responses.
+    // Upstream middleware remains unaffected and continues execution.
+    app.include(rescue::inspect(|error| eprintln!("error: {}", error)));
 
     // Serve any file located in the public dir.
     app.at("/*path").respond(via::get(file_server));
