--- src/error/mod.rs
@@ -1,5 +1,0 @@
-mod error;
-mod server;
-
-pub use error::{BoxError, Error, Message};
-pub(crate) use server::ServerError;
