--- src/middleware/mod.rs
@@ -1,14 +1,0 @@
-pub mod cookie_parser;
-pub mod error_boundary;
-
-pub(crate) mod method;
-
-mod filter;
-mod middleware;
-mod next;
-mod timeout;
-
-pub use filter::{Filter, Predicate, filter};
-pub use middleware::{BoxFuture, Middleware, Result};
-pub use next::Next;
-pub use timeout::{Timeout, timeout};
